Hook up barcode scan to search api
